pyproject.toml
[project]
name = "mistral-hackiterate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "emmuzoo", email = "edison.marcelo.muzo.oyana@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.37.0",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "langchain-mistralai>=1.1.1",
    "mistralai>=1.12.4",
    "pydantic-settings>=2.13.1",
    "pydub>=0.25.1",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "trafilatura>=2.0.0",
]

[build-system]
requires = ["uv_build>=0.9.11,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "sentinel"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
